fix(network): fix concrete file operations

+++ networking/handle_datasets.py
@@ -4,17 +4,17 @@ import pathlib as pl
 
 def load_config(filename):
     """
-
+    Load configuration file for file operations.
     """
     with open(filename, "r") as f:
         return json.load(f)
 
 def download_file(url, destination):
     """
-
+    Open and write dataset.
     """
     response=requests.get(url, stream=True)
-    if response.status_code == 200:
+    if response.status_code == 200 and destination.exists():
         with open(destination, "wb") as f:
             for chunk in response.iter_content(chunk_size=8192):
                 f.write(chunk)
@@ -28,12 +28,15 @@ def print_progress():
     """
 
 def fetch_imdb_dataset():
+    """
+    Orchestrate full file operation.
+    """
     requests_config = load_config('dataset.json')
     folder = requests_config['download_dir']
     for dataset in requests_config['datasets']:
         url = dataset['url']
         filename = dataset['filename']
-        destination = pl.Path(folder) / filename
+        destination = pl.Path(__file__).parent.parent / folder / filename
         download_file(url, destination)
 
 if __name__ == '__main__':
